Add profile page and database table

+++ app/views/users/edit.blade.php
@@ -19,18 +19,13 @@
 			@include('layouts.partials.errors')
 
 		
-			{{ Form::model($user, ['method' => 'patch', 'route' => ['users.update', $user->id]]) }}
+			{{ Form::model($user, ['method' => 'patch', 'route' => ['users.update', $user->username]]) }}
 
 			{{ Form::hidden('id',null) }}
 
 			<div class="form-group">
 				{{ Form::label('username', 'Username:') }}
 				{{ Form::text('username', null, ['class' => 'form-control']) }}
-			</div>
-
-			<div class="form-group">
-				{{ Form::label('email', 'Email:') }}
-				{{ Form::text('email', null, ['class' => 'form-control']) }}
 			</div>
 
 			<div class="form-group">
@@ -44,10 +39,73 @@
 			</div>
 
 			<div class="form-group">
+				{{ Form::label('email', 'Email:') }}
+				{{ Form::email('email', null, ['class' => 'form-control']) }}
+			</div>
+
+
+
+			<div class="form-group">
+				{{ Form::label('tel', 'Teléfono:') }}
+				{{ Form::text('profile[tel]', null, ['class' => 'form-control']) }}
+			</div>
+
+			<div class="form-group">
+				{{ Form::label('cel', 'Celular:') }}
+				{{ Form::text('profile[cel]', null, ['class' => 'form-control']) }}
+			</div>
+
+			<div class="form-group">
+				{{ Form::label('address', 'Dirección:') }}
+				{{ Form::text('profile[address]', null, ['class' => 'form-control']) }}
+			</div>
+
+			<div class="form-group">
+				{{ Form::label('colonia', 'Colonia:') }}
+				{{ Form::text('profile[colonia]', null, ['class' => 'form-control']) }}
+			</div>
+
+			<div class="form-group">
+				{{ Form::label('city', 'Ciudad:') }}
+				{{ Form::text('profile[city]', null, ['class' => 'form-control']) }}
+			</div>
+
+			<div class="form-group">
+				{{ Form::label('state', 'Estado:') }}
+				{{ Form::text('profile[state]', null, ['class' => 'form-control']) }}
+			</div>
+
+			<div class="form-group">
+				{{ Form::label('cp', 'Código Postal:') }}
+				{{ Form::text('profile[cp]', null, ['class' => 'form-control']) }}
+			</div>
+
+			<div class="form-group">
+				{{ Form::label('country', 'País:') }}
+				{{ Form::select('profile[country]', array('MX' => 'México', 'ES' => 'España', 'US' => 'United States', 'CR' => 'Costa Rica'), null, ['class' => 'form-control']) }}
+			</div>
+
+			<div class="form-group">
+				{{ Form::label('birthday', 'Birthday:') }}
+				{{ Form::text('profile[birthday]', null, ['class' => 'form-control', 'placeholder' => 'YYYY-MM-DD']) }}
+			</div>
+
+			<div class="form-group">
+				{{ Form::label('bio', 'Bio:') }}
+				{{ Form::textarea('profile[bio]', null, ['class' => 'form-control']) }}
+			</div>
+
+
+
+
+
+
+			<div class="form-group">
 				{{ Form::label('isactive', 'Currently Active: ') }}
 				{{ Form::checkbox('isactive','1',true) }}
 			</div>
 
+		
 			<div class="form-group">
 				{{ Form::label('isinvisible', 'Hidden: ') }}
 				{{ Form::checkbox('isinvisible','1',false) }}
@@ -59,11 +117,11 @@
 
 			{{ Form::close() }}
 			
-
+		@if(1 == 0)
 			{{ Form::model($user, ['method' => 'delete', 'route' => ['users.destroy', $user->id]]) }}
 				{{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
 			{{ Form::close() }}
-
+		@endif
 
 
 		</div>
